pyproject.toml
[project]
name = "key2midi-pad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pynput", "mido", "python-rtmidi", "pywebview", "pyinstaller", "requests", "packaging"]

[project.scripts]
key2midi-pad = "src.main:main"
